# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ParamUpdateEmail:
    properties:
      email:
        type: string
      verify_code:
        type: string
    type: object
  models.ParamUpdatePassword:
    properties:
      new_password:
        type: string
      new_password_confirm:
        type: string
      password:
        type: string
    type: object
  models.ParamUpdatePhone:
    properties:
      code:
        type: string
      phone:
        type: string
    type: object
  models.ParamUpdateProfile:
    properties:
      city:
        type: string
      introduction:
        type: string
      name:
        type: string
    type: object
  models.ParamVoteData:
    properties:
      direction:
        description: required会把一些零值给看做没有值， 比如0对于int
        enum:
        - 0
        - 1
        - -1
        example: "0"
        type: string
      post_id:
        example: "0"
        type: string
    required:
    - post_id
    type: object
...
host: localhost:9000
info:
  contact:
    email: 1298453249@qq.com
    name: xiaorui zheng
    url: http://www.swagger.io/support
  description: 这是一个热点论坛项目， 能够根据当下热点来向用户展示论坛帖子
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 热点论坛
  version: "1.0"
paths:
...
  /post/vote:
    post:
      consumes:
      - application/json
      description: 对于某个帖子进行投票
      parameters:
      - description: 参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamVoteData'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 对于某个帖子进行投票
      tags:
      - Post
  /user:
    get:
      consumes:
      - application/json
      description: 获得当前用户信息
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获得当前用户信息
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 更改用户个人信息
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamUpdateProfile'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更改用户个人信息
      tags:
      - User
  /user/avatar:
    put:
      consumes:
      - application/json
      description: 更新用户头像
      parameters:
      - description: Avatar file
        in: formData
        name: avatar
        required: true
        type: file
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户头像
      tags:
      - User
  /user/email:
    put:
      consumes:
      - application/json
      description: 更改邮箱
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamUpdateEmail'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更改邮箱
      tags:
      - User
  /user/password:
    put:
      consumes:
      - application/json
      description: 更新用户密码
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamUpdatePassword'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户密码
      tags:
      - User
  /user/phone:
    put:
      consumes:
      - application/json
      description: 更改手机号码
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamUpdatePhone'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更改手机号码
      tags:
      - User
swagger: "2.0"